Split out useActions hook
